Break up elections component
